another try to fix the installation problem -- moved ResponseParameters.hpp to chem lib
